requirements.txt
a2wsgi==1.10.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.5
exceptiongroup==1.2.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==5.13.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
packaging==23.2
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
swagger_ui_bundle==1.1.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
Werkzeug==3.0.1
zipp==3.17.0
